done with beta_lpdf

+++ stan/math/opencl/prim/size.hpp
@@ -14,7 +14,7 @@ namespace math {
  */
 template <typename T,
           require_all_kernel_expressions_and_none_scalar_t<T>* = nullptr>
-int size(const T& a) {
+size_t size(const T& a) {
   return a.rows() * a.cols();
 }
